u: 의존성 방향이 `engine > store`가 되도록 수정
